move utils and models to top-level dir, out of tools dir
